requirements.txt
# Core dependencies
gradio
PyYAML

# Document processing
pdfplumber
PyMuPDF
pdf2image
Pillow

# Table extraction
tabula-py
camelot-py[cv]

# OCR capabilities
pytesseract
easyocr
# paddlepaddle
# paddleocr

# Text processing
python-docx

# LangChain ecosystem
# langchain>=0.1.0
# langchain-community>=0.0.20
# langchain-huggingface>=0.0.1
# langchain-core>=0.1.0

# Vector stores and embeddings  
# chromadb>=0.4.0
# sentence-transformers>=2.2.0

# Ollama integration
# ollama>=0.1.0

# Scientific computing
numpy>=1.24.0

# Image processing
opencv-python>=4.8.0

# Optional: Layout detection
# layoutparser
# detectron2

# Optional: Alternative vector store
# faiss-cpu>=1.7.4

# Development and testing (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0